+++ src/utils/with-two-sources-helper.js
@@ -30,24 +30,25 @@ function withTwoSources(name, mixin /*, options*/) {
 
     _removeSecondary: function() {
       if (this._secondary !== null) {
-        this._secondary.offAny([this._handleSecondaryAny, this]);
+        this._secondary.offAny(this._$handleSecondaryAny);
+        this._$handleSecondaryAny = null;
         this._secondary = null;
       }
     },
 
     _onActivation: function() {
       if (this._secondary !== null) {
-        this._secondary.onAny([this._handleSecondaryAny, this]);
+        this._secondary.onAny(this._$handleSecondaryAny);
       }
       if (this._alive) {
-        this._primary.onAny([this._handlePrimaryAny, this]);
+        this._primary.onAny(this._$handlePrimaryAny);
       }
     },
     _onDeactivation: function() {
       if (this._secondary !== null) {
-        this._secondary.offAny([this._handleSecondaryAny, this]);
+        this._secondary.offAny(this._$handleSecondaryAny);
       }
-      this._primary.offAny([this._handlePrimaryAny, this]);
+      this._primary.offAny(this._$handlePrimaryAny);
     }
   }, mixin || {});
 
@@ -60,6 +61,9 @@ function withTwoSources(name, mixin /*, options*/) {
       this._secondary = secondary;
       this._name = primary._name + '.' + name;
       this._lastSecondary = NOTHING;
+      var $ = this;
+      this._$handleSecondaryAny = function(event) {  $._handleSecondaryAny(event)  }
+      this._$handlePrimaryAny = function(event) {  $._handlePrimaryAny(event)  }
       this._init();
     }
 
@@ -69,6 +73,8 @@ function withTwoSources(name, mixin /*, options*/) {
         this._primary = null;
         this._secondary = null;
         this._lastSecondary = null;
+        this._$handleSecondaryAny = null;
+        this._$handlePrimaryAny = null;
         this._free();
       }
     }, mixin);